refactor: centralize date utilities and remove duplicates
